Add an endpoint that takes in data and pushes it to MinIO

+++ deltafi-action-kit/src/main/java/org/deltafi/actionkit/action/format/FormatResult.java
@@ -54,7 +54,7 @@ public class FormatResult extends Result {
     }
 
     @SuppressWarnings("unused")
-    public void setObjectReference(@NotNull String name, @NotNull String bucket, int size, int offset) {
+    public void setObjectReference(@NotNull String name, @NotNull String bucket, long size, long offset) {
         objectReferenceInput = new ObjectReferenceInput(name, bucket, offset, size);
     }
